Convert "Update TFT display" into a button
* This is a breaking change. Existing automation using the Switch version will need an update to use the button version instead.

This was originally set as a switch, but in reality it behaves as a button, as when the user click it runs an action (in this case, update the TFT display file). Switches should be used only when an entity will remain in one of the two states (on/off), which is not the case here.

This is related to #328.