Reduces unnecessary runs
This adds condition to the automation to preventing it to run when an update to an entity which is not in the current page triggers the automation.
This will reduce the number of parallel instances and with this the number of entries in the logbook. (#304)
Will enable future development to update entity pages without compromising overall performance. (#410)