Reduce number of unnecessary executions
Currently, many triggers will start the automation and then later if a condition is not true (like a page is not visible) the automation won't be doing anything. However, each of those "false" executions will generate an entry at Logbook, which might get spammed for large systems.

Probably solves #1205
Probably helps with #1203

This also reduce the date update from every minute to every hour and handle the situation where multiple panels are updated simultaneously in a event of an automation reload, which could cause performance issues for large systems.