Add support for easily changing backgrounds
Changing the background of the panel was a common request, but due to technical limitations in the Nextion panel's handling of images it was not a straightforward task to do. This commit tries to fix this by introducing 3 changes to make the process more streamlined.

1. Making sure cropped images are always on the same place

This mainly involves the weather icons, as they are in a slightly different place between the Home page and the Weather pages. The changes make sure they are at the same place as the Home page everywhere. On the US version this also requires the temperature bit to move as well on the Weather page slightly.

The other element that needs to be moved was the Fan control on the US version, so it is in the same place as the Cover control

A couple other minor fixes in the HMI files include making sure texts properly have the "Crop Image" setting applied instead of a black background on a few pages, and conversely it should be "Solid Color" on the few pages where the background is fixed black. Although unused the home_smpl page was also fixed up in the US version as it had misaligned boxes.

2. Autogenerate opaque images

Since Nextion doesn't support transparent PNGs they need to be made opaque in pre-processing. To make this task easier there's a new command called `generate_images.sh` in the `ui` folder that will take the transparent images and put the background behind them at the appropriate place (for the weather icons and cover controls).

Note: the transparent images have been generated from the black background ones using imagemagick's remove background feature for the weather icons, and manually makind the images transparent on some other pages (climate control, settings). If the original layered images are available (e.g. in XCF or PSD) I could generate these files more properly

3. Automate TFT generation

The third hurdle is usually the fact that Nextion Editor is a Windows GUI tool without any command line interface to make changes in the HMI files (and TFT compilation) easy. I am trying to address this with a separate project at https://github.com/sztupy/nextion-editor-docker which is a docker container containing Wine, Nextion Editor and AutoHotKey allowing some changes to be done in an automated way. There are currently two scripts that are likely useful:

* Update the version number in the HMI file and generate a TFT
* Replace all images in the HMI file and generate a TFT

The second bit I also made it into a Home Assistant addon, where you can put your background files in the `/config` directory, and run the scripts to generate tft files at the `http://homeassistant.local:8123/local/<name>.tft` location, allowing easy updates afterwards from the ESPHome device page