Change HMI->TXT conversion, including helpers
Changes how we store the Nextion2Text converted files, with both visual and unknown elements enabled, to make it easier to check both code and visual changes between versions.

Also includes a couple helper scripts to allow managing of HMI files easier. They are:

* `.github/workflows/validate_hmi2txt.yml`: makes sure that the HMI and TXT files are in sync
* `hmi/dev/scripts/generate_nextion_txt.sh`: Generates the TXT files from the HMI files. Uses a fork of Nextion2Text that includes support for the "Touch Move" event as well as some changes to report binary data as HEX in the export
* `hmi/dev/scripts/update_nextion_version.sh`: Updates the version number inside the HMI files to the specified one

Example usage of the scripts:

```sh
./hmi/dev/scripts/update_nextion_version.sh 4.4.0.dev10
./hmi/dev/scripts/generate_nextion_txt.sh
```

Note: the Blank HMI file needed updates to move the "version" variable to the leftmost position, the same place as it's on the other HMI files. This is to allow the version update script to find the variable and update it