Add generator for US_land and CJK HMI versions
Technically we only have two different designs: EU and US, and the other 4 are derived from these two with some minor modifications.

This change adds a script that will do these modifications automatically.

Just like with the version update script, this only works on `x86_64` right now, and would preferably need a mid-range PC with enough RAM to run with a high success rate. Takes around 10 minutes to finish. On lower range devices (like an Intel NUC), it might need a couple retries, and more time.

Note: the two fonts directory are the same files as are in the HMI files, but extracted. This makes it easier to import them both manually and via scripts, as you can just "Delete All" then "Import" them in two steps. The old fonts directory was moved to `source`.