Split wakeup and dim, so the watchdog would only force the dim every x:30 seconds
`timer_dim` was used not just to initiate the dimming, but also to actually wake up the panel. However since `timer_dim` was also added to the timer watchdog, it meant the panel was woken up every minute.

This change splits the dimming and the wakeup into two separate scripts, so the dimming can still be part of the watchdog, while the rest of the commands can do both wakeup + dim.