Review `page_changed` and move ´action_button` to script
With the delays implemented to solve the boot issue with ESPHome 2025.8.0, a single button was taking longer to be sent to the display, then when the blueprint sent the action for the next button, it either fails or cancels the previous button, causing that only one button was displayed.
This fix moves the button constructor to a script which will queue the requests and plot all buttons.